+++ browser/i18n/da_DK.js
@@ -221,6 +221,9 @@ module.exports = {
         "Apply": "Anvend",
         "Reset filter": "Nulstil filter",
         "Filter editor": "Filter editor",
+        "Layer is visible": "Laget er synligt",
+        "Parents": "Forældre",
+        "Children": "Børn",
 
         // Base layers
         "Side-by-side mode": "Dobbelt baggrundskort",