+++ browser/i18n/en_US.js
@@ -221,6 +221,9 @@ module.exports = {
         "Apply": "Apply",
         "Reset filter": "Reset filter",
         "Filter editor": "Filter editor",
+        "Layer is visible": "Layer is visible",
+        "Parents": "Parents",
+        "Children": "Children",
 
         // Base layers
         "Side-by-side mode": "Side-by-side mode",